requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
motor==3.6.0
pymongo==4.10.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.2

